Add files via upload
The D'mart Clothing Sale Dashboard encapsulates a wealth of data-driven insights, providing a holistic view of sales performance. By harnessing the power of Power BI and the richness of the dataset, this dashboard equips stakeholders with the tools necessary to make informed decisions, adapt strategies, and drive the success of the clothing sales enterprise.